Fix legend margins on plots 3 and 4
__PROBLEM__

For whatever reason, the default sizing of the legend box when using
the labels from the example caused margin collapse on the right
side -- they look ok on the actual plot, but when written to the png
device the text cramps or overlaps the right border of the graph.

__SOLUTION__

After trying many variations of par values for legend, I ended up
just adding whitespace to the ends of the string label names
themselves.  It's hacky, and I don't love it.  But it works.
No more margin overrun!